Added option to disable dats, cleaned stuff, changed readme to rst

+++ lib/actions/processor.py
@@ -11,26 +11,12 @@ from lib.database.models.datfile import Dat
 class Processor:
     """ Process actions. """
     _previous = None
+    actions = []
+    repo = None
+    file = None
 
     def __init__(self, **kwargs):
         self._previous = None
-        self.repo = 'redump'
-        self.file = '/mnt/i/E/ROMVault/UpdateDats/tmp/redump/dats/bios/Microsoft - Xbox - BIOS Datfile (7) (2010-09-13).dat'
-        self.actions = [
-            {
-                'action': 'LoadDatFile',
-                'class_name': 'lib.repos.redump.redump_dat.RedumpBiosDat'
-            },
-            {
-                'action': 'DeleteOld'
-            },
-            {
-                'action': 'Copy'
-            },
-            {
-                'action': 'SaveToDatabase'
-            }
-        ]
         self.__dict__.update(kwargs)
 
     def process(self):
@@ -62,7 +48,6 @@ class LoadDatFile(Process):
         self._class = locate(self.class_name)
         self._dat = self._class(file=self.file)
         self._dat.load()
-        # self._dat.parse()
         self.database = Dat(repo=self.repo, **self._dat.dict())
         self.output = self.database.dict()
 
@@ -76,7 +61,7 @@ class DeleteOld(Process):
         self.database.load()
         olddat = self.database.dict()
         if 'new_file' in olddat and os.path.exists(olddat['new_file']):
-            os.unlink(self.database.dict()['new_file'])
+            os.unlink(olddat['new_file'])
 
         self.output = self.previous
 
@@ -86,6 +71,7 @@ class Copy(Process):
     destination = None
     file = None
     folder = None
+    database = None
 
     def process(self):
         """ Copy files. """
@@ -93,11 +79,23 @@ class Copy(Process):
             origin = self.file
         filename = os.path.basename(origin)
         self.destination = self.destination if self.destination else self.previous['path']
+
         destination = os.path.join(Settings.ROMVAULT_PATH, self.folder, self.destination, filename)
         os.makedirs(os.path.dirname(destination), exist_ok=True)
-        os.system(f'cp "{origin}" "{destination}"')
         if self.previous:
-            self.previous['new_file'] = destination
+            self.database = Dat(repo=self.previous['repo'], name=self.previous['name'])
+            self.database.load()
+            if self.database.is_enabled():
+                old_file = os.path.basename(self.database.dict()['new_file'])
+                new_file = os.path.basename(destination)
+                if old_file != new_file or Settings.Overwrite or not os.path.exists(destination):
+                    self.previous['new_file'] = destination
+                    os.system(f'cp "{origin}" "{destination}"')
+            else:
+                self.previous['new_file'] = None
+        else:
+            os.system(f'cp "{origin}" "{destination}"')
+
         self.output = self.previous
